Split BattlePopup into a create and edit variant
